+++ CHANGELOG.md
@@ -5,9 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [1.2] - 2022-08-09
 ### Added
 - Pyright static checker
 - Added `/schedule list` and `/schedule remove`
+- Added version info to `/info` and `/help` commands
 
 ### Fixed
 - Many internal typing issues
@@ -45,7 +48,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Initial code base for the bot
 
 
-[Unreleased]: https://github.com/Taaku18/discord-message-scheduler/compare/v1.1.1...HEAD
+[Unreleased]: https://github.com/Taaku18/discord-message-scheduler/compare/v1.2...HEAD
+[1.2]: https://github.com/Taaku18/discord-message-scheduler/compare/v1.1.1...v1.2
 [1.1.1]: https://github.com/Taaku18/discord-message-scheduler/compare/v1.1...v1.1.1
 [1.1]: https://github.com/Taaku18/discord-message-scheduler/compare/v1.0...v1.1
 [1.0]: https://github.com/Taaku18/discord-message-scheduler/releases/tag/v1.0